[Misc] Move jsontree to utils
