+++ vllm/multimodal/cache.py
@@ -7,9 +7,9 @@ from typing import TypeVar, Union
 
 import torch
 
-from vllm.jsontree import json_map_leaves, json_reduce_leaves
 from vllm.logger import init_logger
 from vllm.utils import GiB_bytes, LRUCache
+from vllm.utils.jsontree import json_map_leaves, json_reduce_leaves
 
 from .inputs import MultiModalKwargs, MultiModalKwargsItem, NestedTensors
